feat: implement format handler system with ChatGPT and Claude handlers
