+++ src/loaders/chatgpt.ts
@@ -1,140 +1,24 @@
-import {
-  type ChatGPTConversation,
-  type ChatGPTNode,
-  chatGPTConversationSchema,
-} from "../schemas/chatgpt.js";
+import { defaultRegistry } from "../core/handler-registry.js";
+import { ChatGPTHandler } from "../handlers/chatgpt-handler.js";
 import type { Conversation } from "../types.js";
-import { logLoadingSummary } from "../utils/loader-logger.js";
-import {
-  formatValidationReport,
-  validateWithDetails,
-} from "../utils/schema-validator.js";
 
-export async function loadChatGPT(
+// Register handler if not already registered
+if (!defaultRegistry.hasHandler("chatgpt")) {
+  defaultRegistry.register(new ChatGPTHandler());
+}
+
+async function loadChatGPT(
   data: unknown,
   options: { quiet?: boolean } = {},
 ): Promise<Conversation[]> {
+  // Backward compatibility check
   if (!Array.isArray(data)) {
     throw new Error("ChatGPT export data must be an array");
   }
 
-  const conversations: Conversation[] = [];
-  const validationErrors: string[] = [];
-  const skippedFields = new Set<string>();
-  let successCount = 0;
-
-  for (let i = 0; i < data.length; i++) {
-    const result = validateWithDetails(chatGPTConversationSchema, data[i], {
-      name: `Conversation #${i + 1}`,
-    });
-
-    if (!result.success) {
-      const report = formatValidationReport(result);
-      validationErrors.push(`Conversation #${i + 1}:\n${report}`);
-      continue;
-    }
-
-    if (result.warnings) {
-      // Collect unknown fields
-      for (const warning of result.warnings) {
-        if (warning.unknownFields) {
-          warning.unknownFields.forEach((field) => skippedFields.add(field));
-        }
-      }
-    }
-    successCount++;
-
-    const parsed = result.data as ChatGPTConversation;
-    const messages = extractMessages(
-      parsed.mapping as Record<string, ChatGPTNode>,
-    );
-
-    const date = parsed.create_time
-      ? new Date(parsed.create_time * 1000)
-      : new Date();
-
-    conversations.push({
-      id: parsed.id || Object.keys(parsed.mapping)[0] || "unknown",
-      title: parsed.title || "Untitled Conversation",
-      date,
-      messages,
-    });
+  const handler = defaultRegistry.getById("chatgpt");
+  if (!handler) {
+    throw new Error("ChatGPT handler not found");
   }
-
-  if (validationErrors.length > 0) {
-    throw new Error(
-      `Schema validation error:\n${validationErrors.join("\n\n")}`,
-    );
-  }
-
-  // Display summary information
-  logLoadingSummary({
-    successCount,
-    exportType: "ChatGPT",
-    skippedFields,
-    quiet: options.quiet ?? false,
-  });
-
-  return conversations;
-}
-
-function extractMessages(
-  mapping: Record<string, ChatGPTNode>,
-): Conversation["messages"] {
-  const messages: Conversation["messages"] = [];
-  const rootNodes = Object.values(mapping).filter(
-    (node) =>
-      !Object.values(mapping).some((n) => n.children?.includes(node.id)),
-  );
-
-  for (const root of rootNodes) {
-    traverseNode(root.id, mapping, messages);
-  }
-
-  return messages.filter((m) => m.content);
-}
-
-function traverseNode(
-  nodeId: string,
-  mapping: Record<string, ChatGPTNode>,
-  messages: Conversation["messages"],
-): void {
-  const node = mapping[nodeId];
-  if (!node) return;
-
-  if (node.message?.content?.parts && node.message.content.parts.length > 0) {
-    const contentParts = node.message.content.parts
-      .map((part) => {
-        if (typeof part === "string") {
-          return part;
-        } else if (part && typeof part === "object") {
-          // For objects with content_type and text fields
-          if ("text" in part && typeof part.text === "string") {
-            return part.text;
-          }
-          // Stringify other objects
-          return JSON.stringify(part, null, 2);
-        }
-        return "";
-      })
-      .filter((part) => part !== "");
-
-    if (contentParts.length > 0) {
-      const message: Conversation["messages"][number] = {
-        role: node.message.author.role,
-        content: contentParts.join("\n"),
-      };
-
-      // Add timestamp only if it exists
-      if (node.message.create_time) {
-        message.timestamp = new Date(node.message.create_time * 1000);
-      }
-
-      messages.push(message);
-    }
-  }
-
-  if (node.children && node.children.length > 0 && node.children[0]) {
-    traverseNode(node.children[0], mapping, messages);
-  }
+  return handler.load(data, options);
 }
